# Avoid duplicate dir search