@vdukhovni 2024-04-17
Performance comparison from `x509storeissuer`, with thread counts of 1, 10, 100 and 1000:

Baseline is master without this PR, 

|Threads|Baseline (us)|PR (us)|
|---|---|---|
|   1 | 3107|1459|
|  10 | 8505|2432
| 100 | 156688|2560|
|1000 |1382203|2501|