@vdukhovni 2024-04-18
> thanks for fixing the tests. I think I've hit this bug earlier this week when running tests late in night my time.

Note that the test fixes are a [separate PR](https://github.com/openssl/openssl/pull/24139) that is a merge-base for this one. Once that is merged, this will be rebased to contain only the last commit. The base PR has two approvals, so I hope will be merged shortly.