@levitte 2024-04-18
> I think this is ready for review (see updated comment above). Dropping legacy engine support from the store is I think a separate issue, that can be tackled in another PR... @levitte perhaps?

That would be a breaking change for anyone still using an engine with a scheme implementation.  I believe that the pkcs11 engine is one of those...  no?  Anyway, that would defer such a change to OpenSSL 4.0 at the earliest, and should probably happen in sync with general removal of engine support (which is currently only deprecated).