@levitte 2024-04-18
> IMHO the URI schemes should still be **eagerly** resolved to their lookup methods at the time the URI is added to the store, rather than during the lookup. Are there compelling reasons to defer the binding?

This sounds like a "what function to use" question, really.  We have two functions to add a URI to a OSSL_STORE lookup:

- `X509_LOOKUP_add_store()`, which just adds the URI to an internal list, from which objects will be loaded later.  This is the lazy function.
- `X509_LOOKUP_load_store_ex()` (and the older `X509_LOOKUP_load_store()`), which loads all objects from the given URI immediately.  This is the eager function.

So it seems to me that the choice is left to the user.

But, perhaps you're trying to say that we should use `X509_LOOKUP_load_store_ex()` in `X509_STORE_set_default_paths_ex()`, rather than `X509_LOOKUP_add_store()`?